fix!: switch tye of device ids

+++ src/main/kotlin/com/benruehl/application/services/LinkService.kt
@@ -17,8 +17,8 @@ class LinkService {
         val createdLink = linkRepository.save(
             Link(
                 id = null,
-                sourceDeviceId = link.sourceDeviceId,
-                targetDeviceId = link.targetDeviceId
+                sourceDeviceId = link.sourceDeviceId.toInt(),
+                targetDeviceId = link.targetDeviceId.toInt()
             )
         )
         return QueryLinkResponse(
